+++ include/arch/cluster/k1b-cluster/memory.h
@@ -121,13 +121,13 @@
 		#define K1B_CLUSTER_HYPER_HIGH_END_PHYS  0x00200000 /**< High Hypervisor End  */
 	#endif
 #ifndef _ASM_FILE_
-	extern const paddr_t K1B_CLUSTER_KERNEL_BASE_PHYS;      /**< Kernel Base          */
-	extern const paddr_t K1B_CLUSTER_KERNEL_END_PHYS;       /**< Kernel End           */
-	extern const paddr_t K1B_CLUSTER_KSTACK_BASE_PHYS;      /**< Kernel Stack Base    */
-	extern const paddr_t K1B_CLUSTER_KPOOL_BASE_PHYS;       /**< Kernel Pool Base     */
-	extern const paddr_t K1B_CLUSTER_KPOOL_END_PHYS;        /**< Kernel Pool End      */
-	extern const paddr_t K1B_CLUSTER_USER_BASE_PHYS;        /**< User Base            */
-	extern const paddr_t K1B_CLUSTER_USER_END_PHYS;         /**< User End             */
+	EXTERN const paddr_t K1B_CLUSTER_KERNEL_BASE_PHYS;      /**< Kernel Base          */
+	EXTERN const paddr_t K1B_CLUSTER_KERNEL_END_PHYS;       /**< Kernel End           */
+	EXTERN const paddr_t K1B_CLUSTER_KSTACK_BASE_PHYS;      /**< Kernel Stack Base    */
+	EXTERN const paddr_t K1B_CLUSTER_KPOOL_BASE_PHYS;       /**< Kernel Pool Base     */
+	EXTERN const paddr_t K1B_CLUSTER_KPOOL_END_PHYS;        /**< Kernel Pool End      */
+	EXTERN const paddr_t K1B_CLUSTER_USER_BASE_PHYS;        /**< User Base            */
+	EXTERN const paddr_t K1B_CLUSTER_USER_END_PHYS;         /**< User End             */
 #endif
 	/**@}*/
 
@@ -148,13 +148,13 @@
 	#endif
 	#define K1B_CLUSTER_USTACK_BASE_VIRT K1B_CLUSTER_HYPER_HIGH_BASE_VIRT /**< User Stack */
 #ifndef _ASM_FILE_
-	extern const vaddr_t K1B_CLUSTER_KERNEL_BASE_VIRT;            /**< Kernel Base          */
-	extern const vaddr_t K1B_CLUSTER_KERNEL_END_VIRT;             /**< Kernel End           */
-	extern const vaddr_t K1B_CLUSTER_KSTACK_BASE_VIRT;            /**< Kernel Stack Base    */
-	extern const vaddr_t K1B_CLUSTER_KPOOL_BASE_VIRT;             /**< Kernel Pool Base     */
-	extern const vaddr_t K1B_CLUSTER_KPOOL_END_VIRT;              /**< Kernel Pool End      */
-	extern const vaddr_t K1B_CLUSTER_USER_BASE_VIRT;              /**< User Base            */
-	extern const vaddr_t K1B_CLUSTER_USER_END_VIRT;               /**< User End             */
+	EXTERN const vaddr_t K1B_CLUSTER_KERNEL_BASE_VIRT;            /**< Kernel Base          */
+	EXTERN const vaddr_t K1B_CLUSTER_KERNEL_END_VIRT;             /**< Kernel End           */
+	EXTERN const vaddr_t K1B_CLUSTER_KSTACK_BASE_VIRT;            /**< Kernel Stack Base    */
+	EXTERN const vaddr_t K1B_CLUSTER_KPOOL_BASE_VIRT;             /**< Kernel Pool Base     */
+	EXTERN const vaddr_t K1B_CLUSTER_KPOOL_END_VIRT;              /**< Kernel Pool End      */
+	EXTERN const vaddr_t K1B_CLUSTER_USER_BASE_VIRT;              /**< User Base            */
+	EXTERN const vaddr_t K1B_CLUSTER_USER_END_VIRT;               /**< User End             */
 #endif
 	#define K1B_CLUSTER_USER_BASE_VIRT 0x80000000                 /**< User Base            */
 	#define K1B_CLUSTER_USER_END_VIRT  0xf0000000                 /**< User End             */
@@ -167,9 +167,69 @@
 	/**
 	 * @brief Initializes the Memory Interface.
 	 */
-	extern void k1b_cluster_mem_setup(void);
+	EXTERN void k1b_cluster_mem_setup(void);
 
 #endif /* __NANVIX_HAL */
+
+	/**
+	 * @brief Writes a TLB entry.
+	 *
+	 * @param vaddr      Target virtual address.
+	 * @param paddr      Target physical address.
+	 * @param shift      Page shift.
+	 * @param way        Target set-associative way.
+	 * @param protection Protection attributes.
+	 */
+	EXTERN int k1b_tlb_write(
+			vaddr_t vaddr,
+			paddr_t paddr,
+			unsigned shift,
+			unsigned way,
+			unsigned protection
+	);
+
+	/**
+	 * @brief Invalidates a TLB entry.
+	 *
+	 * @param vaddr Target virtual address.
+	 * @param shift Page shift.
+	 * @param way   Target set-associative way.
+	 */
+	EXTERN int k1b_tlb_inval(vaddr_t vaddr, unsigned shift, unsigned way);
+
+	/**
+	 * @brief Lookups a TLB entry by virtual address.
+	 *
+	 * @param vaddr Target virtual address.
+	 *
+	 * @returns Upon successful completion, a pointer to the TLB entry
+	 * that matches the virtual address @p vaddr is returned. If no
+	 * entry that meets this criteria is found, @p NULL is returned.
+	 */
+	EXTERN const struct tlbe *k1b_tlb_lookup_vaddr(vaddr_t vaddr);
+
+	/**
+	 * @brief Lookups a TLB entry by physical address.
+	 *
+	 * @param paddr Target physical address.
+	 *
+	 * @returns Upon successful completion, a pointer to the TLB entry
+	 * that matches the physical address @p paddr is returned. If no
+	 * entry that meets this criteria is found, @p NULL is returned.
+	 */
+	EXTERN const struct tlbe *k1b_tlb_lookup_paddr(paddr_t paddr);
+
+	/**
+	 * @brief Flushes the TLB.
+	 */
+	EXTERN int k1b_tlb_flush(void);
+
+	/**
+	 * @brief Dumps a TLB entry.
+	 *
+	 * @param idx Index of target entry in the TLB.
+	 */
+	EXTERN void k1b_tlbe_dump(int idx);
 
 #endif /* _ASM_FILE_ */
 
@@ -199,6 +259,77 @@
 	#define KBASE_VIRT   K1B_CLUSTER_KERNEL_BASE_VIRT /**< @see K1B_CLUSTER_KERNEL_BASE_VIRT  */
 	#define KPOOL_VIRT   K1B_CLUSTER_KPOOL_BASE_VIRT  /**< @see K1B_CLUSTER_KPOOL_BASE_VIRT   */
 	/**@}*/
+
+#ifndef _ASM_FILE_
+
+	/**
+	 * @brief Provided Interface
+	 */
+	/**@{*/
+	#define __tlb_lookup_vaddr_fn /**< tlbe_lookup_vaddr() */
+	#define __tlb_lookup_paddr_fn /**< tlbe_lookup()       */
+	#define __tlb_write_fn        /**< tlb_write()         */
+	#define __tlb_inval_fn        /**< tlb_inval()         */
+	#define __tlb_flush_fn        /**< tlb_flush()         */
+	/**@}*/
+
+	/**
+	 * @see k1b_tlb_lookup_vaddr().
+	 */
+	static inline const struct tlbe *tlb_lookup_vaddr(int tlb_type, vaddr_t vaddr)
+	{
+		/* Invalid TLB type. */
+		if ((tlb_type != K1B_TLB_INSTRUCTION) && (tlb_type != K1B_TLB_DATA))
+			return (NULL);
+
+		return (k1b_tlb_lookup_vaddr(vaddr));
+	}
+
+	/**
+	 * @see k1b_tlb_lookup_paddr().
+	 */
+	static inline const struct tlbe *tlb_lookup_paddr(int tlb_type, paddr_t paddr)
+	{
+		/* Invalid TLB type. */
+		if ((tlb_type != K1B_TLB_INSTRUCTION) && (tlb_type != K1B_TLB_DATA))
+			return (NULL);
+
+		return (k1b_tlb_lookup_paddr(paddr));
+	}
+
+	/**
+	 * @see k1b_tlb_write()
+	 */
+	static inline int tlb_write(int tlb_type, vaddr_t vaddr, paddr_t paddr)
+	{
+		/* Invalid TLB type. */
+		if ((tlb_type != K1B_TLB_INSTRUCTION) && (tlb_type != K1B_TLB_DATA))
+			return (-EINVAL);
+
+		return (k1b_tlb_write(vaddr, paddr, 12, 0, K1B_TLBE_PROT_RW));
+	}
+
+	/**
+	 * @see k1b_tlb_inval()
+	 */
+	static inline int tlb_inval(int tlb_type, vaddr_t vaddr)
+	{
+		/* Invalid TLB type. */
+		if ((tlb_type != K1B_TLB_INSTRUCTION) && (tlb_type != K1B_TLB_DATA))
+			return (-EINVAL);
+
+		return (k1b_tlb_inval(vaddr, 12, 0));
+	}
+
+	/**
+	 * @see k1b_tlb_flush().
+	 */
+	static inline int tlb_flush(void)
+	{
+		return (k1b_tlb_flush());
+	}
+
+#endif /* _ASM_FILE_ */
 
 /**@endcond*/
